refactor: decompose PlaylistSelection into focused sub-components
- Extract LibraryControls: search input, sort dropdown, clear button (drawer + non-drawer variants)
- Extract LibraryStatusContent: loading skeletons, unauthenticated prompt, error alert
- Extract LibraryMainContent: tabs bar, FilterChipRow, PlaylistGrid, AlbumGrid, drawer bottom controls
- index.tsx shrinks from 576 → 310 lines; each new file has a single clear responsibility
